Make wiki folder instead of just having a README.md
Seem to be making some people call a 'second brain'. Might use the style of Microsoft Azure DevOps because it is familiar. Do not like having a single attachments folder because I would prefer to use images for a topic in the folder that houses that topic. When the topic is obsolete, delete the whole folder without tracking down attachments elsewhere.